Switch ActionDialog to typescript
