+++ src/ActionDialog/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./ActionDialog";
+export { ActionDialogProps } from "./ActionDialog.types";
